feat: add KeyPair-to-JWK conversion (#3770)
* feat: add conversion from KeyPair to JWK

* cleaned up a dependency

* renamed to CryptConverter, moved to crypto-core

* pr remarks

* merged KeyFactory and CryptoConverter

* renamed to crypto-common
