+++ extensions/common/iam/identity-trust/build.gradle.kts
@@ -19,7 +19,9 @@ plugins {
 
 dependencies {
     api(project(":extensions:common:iam:identity-trust:identity-trust-core"))
+    api(project(":extensions:common:iam:identity-trust:identity-trust-issuers-configuration"))
     api(project(":extensions:common:iam:identity-trust:identity-trust-service"))
+    api(project(":extensions:common:iam:identity-trust:identity-trust-sts"))
     api(project(":extensions:common:iam:identity-trust:identity-trust-transform"))
 }
