+++ extensions/common/iam/identity-trust/identity-trust-core/build.gradle.kts
@@ -12,7 +12,7 @@ dependencies {
     implementation(project(":extensions:common:crypto:jws2020"))
     implementation(project(":extensions:common:crypto:jwt-verifiable-credentials"))
     implementation(project(":extensions:common:crypto:ldp-verifiable-credentials"))
-    implementation(project(":extensions:common:crypto:crypto-core"))
+    implementation(project(":extensions:common:crypto:crypto-common"))
     implementation(project(":extensions:common:iam:identity-trust:identity-trust-service"))
     implementation(project(":extensions:common:iam:identity-trust:identity-trust-sts:identity-trust-sts-embedded"))
     implementation(libs.nimbus.jwt)
