+++ spi/common/jwt-spi/build.gradle.kts
@@ -20,12 +20,6 @@ plugins {
 dependencies {
     api(project(":spi:common:core-spi"))
     api(project(":spi:common:token-spi"))
-    api(libs.nimbus.jwt)
-    runtimeOnly(libs.tink) // for EdDSA/Ed25519
-
-
-    testImplementation(libs.bouncyCastle.bcprovJdk18on) // for EdDSA/Ed25519
-
 }
